Fixed failing Border render tests.
They were failing because passing a Pango.Context to FormattedTextImpl
was a hack. Create a pango context on startup that can be used for all
FormattedTextImpls. Fix the expected test output for tests containing
text as the expected output came from the Direct2D renderer where text
rendering is slightly different.