Revert "Only send BindingNotifications on error."
This reverts commit 57e646583fcd540a9b2795ee5e07700e241ea4f7.

For TextBox there can be 2 bindings to Text: one which is the binding to the
view model which should have data validation enabled, and another binding to
the TextPresenter in the template which should not have data validation
enabled, or it will override the view model data validation. For this we need
to be able to distinguish between the two and so bindings with data validation
enabled need to always send BindingNotifications.

Conflicts:
	src/Avalonia.Base/AvaloniaObject.cs
	tests/Avalonia.Markup.Xaml.UnitTests/Data/BindingTests_Validation.cs