Use a different method of displaying TabControl.
Make its ItemContainerGenerator generate `TabItem`s, and display these
in an `ItemsPresenter` in its template. The `Carousel` still uses
`MemberSelector` to select the `Content` property of the `TabItems`.
This makes `TabItem.IsEnabled` work. The downside is that its completely
separate from `TabStrip` so `TabItem` and `TabStripItem` need to be
templated completely separately.