Refactor Window.SizeToContent.
For auto-sizing windows, previously we were relying on the fact that when a window was resized it would immediately fire a `Resized` event. We used the `BeginAutoSizing` method as a flag to say "we're auto-sizing now, any change in size isn't due to the user". When a resize event came in outside of a `BeginAutoSizing` block, we set `SizeToContent = Manual`. On X11 the resized event will not be fired immediately, it will be fired sometime later, causing auto-sizing to be broken there.

Instead change the algorithm to assume that the user has resized the window if a resize event comes in for a size other than the requested size. For this to work, we also need to know the min client size for the window, because if we request less than that, we don't get the size we requested.