Fix window resizing jitter.
As described in https://github.com/zkSNACKs/WalletWasabi/issues/1381, when `HasSystemDecorations == false`, resizing a window using the top and/or left handle caused a "jitter" effect.

This was caused by our handling of `WM_NCCALCSIZE`. To fix it, defer to the default handling of `WM_NCCALCSIZE` and then undo the change made to the rectangle top to display a title bar as described at https://stackoverflow.com/a/2135120/6448.