Fix KeyNotFound exception in Animatable.
When transitions are replaced, add the new transitions before removing the old transitions, so that when the old transition being disposed causes the value to change, there is a corresponding entry in `_transitionStates`.

Also search the `Transitions` collection from last to first to find a matching transition, so that later added transitions have priority.

Fixes #4059