(This is my first PR with you guys, so, if you would give me a little leeway.) I am proposing in this PR the ability to Disable and Enable the Close traffic light on the Mac in Avalonia windows. Currently, you guys do have an EnableClose property on the Window control, but I could not find where on earth you implement it. I did find deep in your code you are doing what I found on Microsoft's sites: EnableMenuItem Win32 calls to implement the DisableCloseButton/EnableCloseButton for Windows. I decided to dig in and find out where and how to do it on Mac.
There are those in my company that brought up the valid point, when I had 'EnableCloseButton' and 'DisableCloseButton' as the methods, and they said why not just EnableClose.
1. This may imply, as they questioned, that it would mean also the Alt-F4/Cmd-W keystrokes and the upper-left (Windows) menu close and Dock right-click -> Quit (OSX) disabling as well.

I would like to propose, if you could, take this Mac implementation (in fact, or as suggestion), to have the feature on OSX, and, I think a micro and macro level of control for the traffic lights on OSX and on Windows be just button, button | menu | keystroke.

Humbly submitted.

PS. In Sandbox is a PoC showing the Disable/Enable on Mac.
If I found how you were making it the EnableClose property in Windows (from low-level win32 call), I would have liked to replicate for OSX. I.e., as property setter, rather than method calls.