win32: Don't raise TextInput event when KeyDown was handled.
On win32, returning handled from `WM_KEYDOWN` doesn't automatically prevent a `WM_CHAR` message, resulting in #5849.

The `TranslateMessage` call is responsible for translating the `WM_KEYDOWN` event to a `WM_CHAR` event, so one needs to "simply" not call this function if the key down event was handled. However there's a problem: `TranslateMessage` is called before the `DispatchMessage` call which dispatches the message to the `WindowImpl.

To get around this, add a `PreProcessMessage` which can be used to handle key down messages outside of the `DispatchMessage` call. This solution broadly comes from WPF, though there the solution is a bit more complicated (it's hooked into using `ComponentDispatcher` events).

Fixes #5849