win32: Don't raise TextInput event when KeyDown was handled.
On win32, returning handled from `WM_KEYDOWN` doesn't automatically prevent a `WM_CHAR` message, resulting in #5849.

Fix this by setting a flag after `WM_KEYDOWN` is handled which will make the following `WM_CHAR` message be ignored.

This is the second attempt at fixing this after it was decided that https://github.com/AvaloniaUI/Avalonia/pull/7351 was not the right approach.

Fixes #5849