Don't multiply logical scroll by ScrollSize.
`ScrollSize` is intended to be the amount scrolled when clicking the up/down line scroll buttons. The `logicalScrollItemSize` calculation should have already taken care of calculating this value for scroll gestures (this wasn't noticed before because the only place where it was checked was with `ItemsPresenter` which has a `ScrollSize.Height` of 1.