Added SetCurrentValue, a prerequisite for #9944 This is a direct analogue to WPF's DependencyObject.SetCurrentValue method. It sets the effective value of the property without detaching any bindings or style setters applied to it.
The existing SetValue method is sufficient when the property's value has local priority. If the priority is anything else, or unknown, SetCurrentValue should be used instead. In particular, it should be used when a control is setting its own StyledProperty values.

This commit only calls SetCurrentValue in one place: TemplateBinding. More uses will be added in future commits.