Bump collection change event priority.
Post collection changes to UI thread with higher priority than `Layout` so that all collection changes are processed before the next layout pass.

`VirtualizingStackPanel` maintains its own mapping of Items -> Containers, and updates that mapping when a `CollectionChanged` event is received, but it also expects that mapping to be up-to-date when a layout is performed.

Fixes #10399