Show popups at touch pointer position.
Hackfix for #10914 on stable branch. Store last pointer position in `TopLevel` for both mouse and touch and use this position in `PopupPositionerExtensions`. To do this added a new interface as adding to the `IInputRoot` interface would be a breaking change. Ideally this interface would be internal but we don't expose internals between `Avalonia.Input` and Avalonia.Control` in 0.10.x so this isn't possible.

Fixes #10914