Checking _operation for null isn't enough.
Instead check if `UpdateCount` is not 0 to allow recursive setting of `Source`.

I have a vague feeling that this isn't actually the correct solution, and that more issues are going to arise around this scenario but for now I'm unable to create a failing test so let's go with the simple solution for now.

Fixes #11617