Fix controls not showing up in accessibility when made visibile (#14424)
* Add failing test for control visibility changing.

There was already a test for when visibility moves from visible to invisible but we were missing one for the other way around.

* Create peer even for invisible controls.

When an invisible control is encountered, unless a peer is created for it, there is nothing to listen for the `IsVisible` property changing to `true`. Make sure we create a peer even for invisible controls; we just don't add them to the child collection.