Fix: Enable seamless forward and backward transition in VirtualizingCarouselPanel (#14125)
* Update VirtualizingCarouselPanel.cs

Changes the calculation of the "forward" flag for the page transition to enable seamless forward and backward transitions.

* Update VirtualizingCarouselPanel.cs

Added Check for more than 2 items in carousel as there is actually no way to determine the correct transition with only 2 items. If there are only 2 items the transition behaviour is not changed.

* Update VirtualizingCarouselPanelTests.cs

Added Unit Tests for cycling through list forward and backward