Fix AlwaysSelected in conjuction with SelectedItem/SelectedIndex binding (#14107)
* Added failing tests for #12733.

* Clear SkipLostSelection on batch update start.

If `Source` is changed during a collection update, then the `Clear()` operation will not be committed immediately due to `_isSourceCollectionChanging` being set. In this case, `update.Operation` will still have `SkipLostSelection == true`, meaning that `LostSelection` will not be raised, causing #12733. Clear the flag manually each time `BeginBatchUpdate` is called to avoid this.

Fixes #12733

---------

Co-authored-by: Max Katz <maxkatz6@outlook.com>