Added IReflectableType dynamic property IsReadOnly checking for DataGrid item
In GetPropertyIsReadOnly method added check whether the item type is inherited from IReflectableType.
If the element type is inherited from IReflectableType, we search for the dynamic property by name in the first element of the DataSource collection.
If a dynamic property is found, we check it for IsReadOnly, search and check for regular properties.
Otherwise, search and check for regular properties.