Make popup focus stealing configurable. (#16642)
* Show Avalonia context menu and toolip...

In native text box in integration test app. Only implemented for win32 right now.

This is to test the two popup behaviors required for native controls:

- The context menu needs focus to be transferred to Avalonia
- The ToolTip must not transfer focus to Avalonia

* Added Popup.TakesFocusFromNativeControl.

By default, if a popup is shown when a native control is focused, focus is transferred back to Avalonia in order for the popup to receive input. If this property is set to false, then the shown popup will not receive input until it receives an interaction which explicitly focuses the popup, such as a mouse click.

The effect of this property can be seen in the Embedding tag of the IntegrationTestApp: hovering over the native text box shows an Avalonia `ToolTip` which does not steal focus from the native text box. Right-clicking to open an Avalonia `ContextMenu` does steal focus so the menu items can be selected using the arrow keys.

Currently only implemented on a win32.

* Show tooltip and context menu on macOS.

* Implement TakeFocus on macOS.

* Add integration tests.

Only tested on win32 so far.

* Integration tests won't work on macOS.

As can be expected at this point, really.

* Update API diff.