Make `CharacterReader` internal.
And by extension, make the static classes which define extension methods on it internal.

To do this, I had to make Avalonia.Base's internals visible to Avalonia.Designer.HostApp, as that has a dependency on it, and exclude `StringCompatibilityExtensions` from there as the one from Avalonia.Base is now visible.